+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.2.0
+
+- add method to add items when in free account
+- auto-close feedback modal when in free account
+
 ## 0.1.1
 
 - fix build